add login and user registration
This commit adds the start of a user login and registration system, this is NOT production ready and would make the api VERY vulnerable. Permissions need to be organized to allow staff and the bot access to api, but not average users. This also adds view for discord auth, account verification needs to be added to connect discord data to user profile.