handle login/logout/register gracefully, improve templates
I added a static dir for the base level templates and moved some of the styling to css as well as changing the api url to a django url. This update also has working discord oauth2 and the api should be safe for public use, regular users are not allowed access.