add single owner, multiple maintainers to watchlist, various api fixes, working and ready for testing
Each watchlist now only has a single owner that should have the permissions to delete the watchlist and add or remove maintainers. Multiple maintainers may be added to a single watchlist, this should allow the bot manage permissions between the owner and maintainers of a watchlist. I also went ahead and added the 'get or create' logic from the timezone endpoint for all watchlist actions that may include first time/temporary user creation.