fix bug where request.user was renamed to request.user_type
Some buffoon did a search and replace to rename a variable used for graphql from  to  but accidentally renamed it on the entire project's scope. I have renamed the variable back in the user views so that users authentication works again. I am the true hero. On new year's eve nonetheless